R26 Bug修复(6项) + tr()合规 + 防御性null检查 + 代码质量
P1修复:
- FrameVisualEditor: onAddField()的typeCombo lambda不再捕获row索引
  改为动态查找cellWidget所在行，解决removeRow/drag-drop后行号失效导致访问错误行或越界

P2修复:
- ToastWidget::show(): 首次连接parent->destroyed信号，避免重复toast时累积冗余连接
- OtaWidget::transferCompleted信号: int改为qint64(elapsed/size)，消除static_cast<int>截断
- DataStatistics::formatRate(): 4处QString改为tr()，支持i18n
- ConnectionControllerReconnect::calcBackoffInterval(): qint64中间变量防止int溢出
- TerminalLayoutManager: 3处applyLayout方法增加qobject_cast null检查

代码质量:
- 常量提取和代码风格统一(BackgroundSettingsPopup, MainWindow, SendController等)

评分: 76 + 1 = 77
变更: 16 files, 93 insertions, 46 deletions

+++ src/ota/OtaWidget.h
@@ -67,7 +67,7 @@ signals:
     void transferStarted(const QString& filename);
 
     /** @brief 传输完成信号 @param filename 文件名 @param elapsed 耗时(ms) @param size 文件大小(字节) */
-    void transferCompleted(const QString& filename, int elapsed, int size);
+    void transferCompleted(const QString& filename, qint64 elapsed, qint64 size);
 
     /** @brief 传输失败信号 @param filename 文件名 @param error 错误原因 */
     void transferFailed(const QString& filename, const QString& error);
